+++ src/main/java/kitchenpos/domain/repository/OrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.repository;
 
+import kitchenpos.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
